requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
requests
PyYAML
supabase
pillow
pytesseract
openai
